Async Generator

+++ python_async_comprehension/0-async_generator.py
@@ -6,7 +6,7 @@ import random
 from typing import AsyncGenerator
 
 
-async def async_generator() -> AsyncGenerator[int, None]:
+async def async_generator() -> AsyncGenerator[float, None]:
     """
     Async Generator that yields random numbers
     between 0 and 10 every second for 10 seconds.
